Add grammar parsing to epo-eng parser
The markup of this dictionary is irregular. The original dictionary
needed to be patched in order to allow a consistant parsing. It has only
been modified, if there was an obvious issue and if it didn't require
too many modifications in the original source. The result is a
realtively robust parser, which deals with the patched, but not yet with
the original version.
It currently only generates the **body** of the dictionary.