ding2tei: add deu-eng specific parser class
Each language supported by the ding2tei importer script should have its
own importer subclass, so that the wild-guessing from the ding plain
text format can be as accurate as possible.
The tokenizer can now parse all expressions in [] {} and abbreviations
in /XYZ/, it extracts grammatical info and it builds a form / sense
structure. It still does a lot of mistakes and a number of things aren't
working yet.