Properly format paragraphs of the <availability> header tag
It doesn't look nice when mashing text into a single paragraph.
Individual paragraphs are now reformatted to fit onto a line with 78
characters, but each paragraph has a paragraph indentation upfront to
structure the text. This works fine for availability. For the noteStmt
node, this is trickier, because there are different tags involved and
hence this simple fix doesn't work.