Remove broken and outdated scripts under testing/
Those scripts (and make targets) started a local dict daemon to query
all words from the dictionary index, to make sure that no headword was
forgotten in the conversion process. This was relevant when we had
scripts to generate the index, but since that's done by dictfmt these
days, these scripts were obsolete.