Rewrite teiaddphonetics (Closes #25)
This is a complete rewrite of the former teiaddphonetics script. The
former version was falling short in quoting and hence lead to incorrect
associations between the headword and the IPA transcription.
It also would consume >  16G of RAM for a 400M dictionary, the new
version takes only a fraction. A similar improvement has been done for
the execution performance. The quoting problem is solved as well.
A plus in the new version is that it helps to identify the causing
punctuation, since punctuation leads to line breaks in the eSpeakNG
output. Since the communication with eSpeakNg is line-based, this can
break the transcription.