teiaddphonetics: add escaping characters; check for existing pronunciation info
This commit adds a few more characters to the exception list to avoid
breaking the eSpeakNG transcription. It also fixes the error output.
In addition, each entry is checked wheter a `<pron` element exists and
if so, no transcription takes place. Please note that this would fail if
multiple orth and only one pron element are present.